# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API",
        "description": "Documentation de l'API avec Swagger",
        "contact": {
            "email": "contact@exemple.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/motorcycle-brands": {
            "get": {
                "tags": [
                    "Motorcycle Brands"
                ],
                "summary": "Get all motorcycle brands",
                "operationId": "ecf4810f772f5ef0f46e563ff8ea5a3d",
                "responses": {
                    "200": {
                        "description": "List of motorcycle brands"
                    }
                }
            },
            "post": {
                "tags": [
                    "Motorcycle Brands"
                ],
                "summary": "Create a new motorcycle brand",
                "operationId": "ca62f9a53aae429317a0d555db0db234",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Honda"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Motorcycle brand created"
                    }
                }
            }
        },
        "/api/motorcycle-brands/{id}": {
            "get": {
                "tags": [
                    "Motorcycle Brands"
                ],
                "summary": "Get a motorcycle brand by ID",
                "operationId": "4353d4620781aea0f89414d16ba72e0f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Motorcycle brand data"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Motorcycle Brands"
                ],
                "summary": "Update a motorcycle brand",
                "operationId": "07d7155dc636ccf7569dd392a5d4e404",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Yamaha"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Motorcycle brand updated"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Motorcycle Brands"
                ],
                "summary": "Delete a motorcycle brand",
                "operationId": "bb9a49f49175528d85d778649c2fc2f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/motorcycles": {
            "get": {
                "tags": [
                    "Motorcycles"
                ],
                "summary": "Get all motorcycles",
                "operationId": "c0f4e4f3a808887d946e4cc85a622b96",
                "responses": {
                    "200": {
                        "description": "List of motorcycles"
                    }
                }
            },
            "post": {
                "tags": [
                    "Motorcycles"
                ],
                "summary": "Create a new motorcycle",
                "operationId": "94a6589d128340e0e630959bfce11074",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "brand_id",
                                    "model_id",
                                    "year_id",
                                    "color"
                                ],
                                "properties": {
                                    "brand_id": {
                                        "type": "integer"
                                    },
                                    "model_id": {
                                        "type": "integer"
                                    },
                                    "year_id": {
                                        "type": "integer"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "Red"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Motorcycle created"
                    }
                }
            }
        },
        "/api/motorcycles/{id}": {
            "get": {
                "tags": [
                    "Motorcycles"
                ],
                "summary": "Get a motorcycle by ID",
                "operationId": "edbfbe97c12800775c8e6f5479e9fbff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Motorcycle data"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Motorcycles"
                ],
                "summary": "Update a motorcycle",
                "operationId": "51d631f35a9d36781f427b8dee84618e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "brand_id",
                                    "model_id",
                                    "year_id",
                                    "color"
                                ],
                                "properties": {
                                    "brand_id": {
                                        "type": "integer"
                                    },
                                    "model_id": {
                                        "type": "integer"
                                    },
                                    "year_id": {
                                        "type": "integer"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "Black"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Motorcycle updated"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Motorcycles"
                ],
                "summary": "Delete a motorcycle",
                "operationId": "380bb8e8e4e201f4304b027a4c56adcc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
...
}